Implement trainer testing and fix up structure.
