+++ CI/espresso_tests/integration_tests/test_ensemble_deployment.py
@@ -141,7 +141,7 @@ class TestEnsembleTraining(ut.TestCase):
                 actions=actions,
             )
 
-            rl_trainer = srl.gyms.Gym(
+            rl_trainer = srl.trainers.ContinuousTrainer(
                 [protocol],
                 loss,
             )
